airflow/requirements.txt
apache-airflow[postgres]==2.8.1
pyspark==3.5.0
pandas==2.0.3
psycopg2-binary==2.9.7